merge IEC104 Branch with Development Branch

* commit '320e5c706ced75138a73a5b27eb345bbcec831d3':
  [General] Minor Adjustment
  [Feature] Report Feature Complete.
  [Feature] Report Feature Complete.
  [General] Controller & Script Cleanup
  [Feature] DO Report Function
  [General] Modify Database
  [General] Implemented DO2 Test Case. Invalid negative response
  [General] InitIEC104OutputConnection
  [General] Output Hex Digits Success
  [General] Add DO Receiver for IEC-104 Command. Print console messages to debuglog

+++ src/script/stopIEC104Simulator.java
@@ -15,10 +15,20 @@ public class stopIEC104Simulator implements InterfaceScript {
             IEC104InitConnection.process.destroy();
         }
         Thread.sleep(1000);
-        if (IEC104InitConnection.process.isAlive()) {
+        if (IEC104InitConnection.process != null && IEC104InitConnection.process.isAlive()) {
             IEC104InitConnection.process.destroyForcibly();
         }
         IEC104InitConnection.process = null;
+
+        if (IEC104InitOutputConnection.process != null && IEC104InitOutputConnection.process.isAlive()) {
+            IEC104InitOutputConnection.process.destroy();
+            IEC104InitOutputConnection.br.close();
+        }
+        Thread.sleep(1000);
+        if (IEC104InitOutputConnection.process != null && IEC104InitOutputConnection.process.isAlive()) {
+            IEC104InitOutputConnection.process.destroyForcibly();
+        }
+        IEC104InitOutputConnection.process = null;
         return null;
     }
